Add key derivation to the restricted kernel
