+++ oak_restricted_kernel/src/lib.rs
@@ -38,7 +38,6 @@
 
 mod acpi;
 mod args;
-pub mod attestation;
 mod avx;
 mod boot;
 mod descriptors;
@@ -56,6 +55,7 @@ pub mod shutdown;
 #[cfg(feature = "simple_io_channel")]
 mod simpleio;
 mod snp;
+pub mod snp_guest;
 mod syscall;
 #[cfg(feature = "vsock_channel")]
 mod virtio;
@@ -264,9 +264,12 @@ pub fn start_kernel(info: &BootParams) -> ! {
         // For now we just generate a sample attestation report and log the value.
         // TODO(#2842): Use attestation report in attestation behaviour.
         let report =
-            attestation::get_attestation([42; 64]).expect("couldn't generate attestation report");
+            snp_guest::get_attestation([42; 64]).expect("couldn't generate attestation report");
         info!("Attestation: {:?}", report);
         report.validate().expect("attestation report is invalid");
+
+        let key = snp_guest::get_derived_key().expect("couldn't derive key");
+        info!("Derived key: {:?}", key);
     }
 
     let mut channel = get_channel(
